Moved SoundPool to new SfxPlayer class
